Check dir with OpenCL kernels exists before trying to copy.

+++ program/validation-armcl-opencl-arm64-v8a/ck-run-validation.sh
@@ -2,15 +2,22 @@
 
 # Collective Knowledge (program)
 #
-# See CK LICENSE.txt for licensing details
-# See CK COPYRIGHT.txt for copyright details
+# See CK LICENSE.txt for licensing details.
+# See CK COPYRIGHT.txt for copyright details.
 #
-# Developer: Grigori Fursin, Grigori.Fursin@cTuning.org
-#            Flavio Vella, dividiti
+# Developers:
+# - Grigori Fursin, Grigori.Fursin@cTuning.org
+# - Flavio Vella, flavio@dividiti.com
+# - Anton Lokhmotov, anton@dividiti.com
 #
 
+if [ ! -d "${CK_ENV_LIB_ARMCL_CL_KERNELS}" ]; then
+  echo "ERROR: Directory \`${CK_ENV_LIB_ARMCL_CL_KERNELS}\' with OpenCL kernels does not exist!"
+  exit 1
+fi
+
 echo ""
-echo "Copying CL kernels ..."
+echo "Copying OpenCL kernels ..."
 echo ""
 
 rm -rf cl_kernels
@@ -18,7 +25,7 @@ mkdir cl_kernels
 cp ${CK_ENV_LIB_ARMCL_CL_KERNELS}/* cl_kernels
 
 echo ""
-echo "Running original ARMCL validation ..."
+echo "Running original ArmCL validation ..."
 echo ""
 
 ./arm_compute_validation $1 --filter=${FILTER}
